+++ CHANGES.md
@@ -1,4 +1,5 @@
 #### [unreleased]
+* add new WP-Cron task to run concurrently to `wp_version_check` so that Git Updater managed plugins and themes can take advantage of auto updating
 
 #### 10.3.4 / 2021-06-22
 * refactor `get_repo_requirements()`